Update background image
Set to default image. The background image's sky color is the same
as the background color is CSS.

Having any other image creates problem in different resolutions.
- The image with multiple backgrounds becomes shorter and the
background is difficult to match the different black and white images

- Replacing with a different image and fixing the background color
in CSS also creates an issue as the image seems irregular when it starts
repeating while scrolling.